Add address_scope_db to neutron/models/head.py
The head.py file ensures that all relevant models are
imported when ModelsMigrationSync tests are run in
neutron/tests/functional/db/test_migrations.   This is
one of the models that needs to be present else the
"address_scopes" table comes out as missing.   The
error is usually concealed by the fact that the module
was imported as a side effect of other tests but
is revealed if one runs test_migrations.py alone.

Change-Id: I954d8385c479de663ef1a5ef439866d2d927bcf3