Support qos rules and fields parameters in GET requests
Previously we didn't load the rules into policy object. This patch adds
loading the rules and defines bandwidth_limit_rules as a policy
resource in a single transaction. As a part of moving towards usage of
single transaction, create() and update() of rule were modified
accordingly.
Finally, we support types in GET requests in this patch.

API tests will follow in different patch.

Change-Id: I25c72aae74469b687766754bbeb749dfd1b8867c