Avoid printing log options multiple times
This patch ensures log_opt_values is invoked only once, thus
avoiding annoying multiple dumps of options values.

Note that is rpc_workers is greater than 0 a second dump will
be logged. This patch does not address this as REST and RPC
servers are going to be separated as a part of the feature/pecan
work.

Closes-Bug: #1477975

Change-Id: Ia5dcb609241de6ad30d9831c5fb98a9e2be6ad7f