objects.qos.policy: provide rules field, not type specific
It should be forbidden to have multiple rules of the same type attached
to a policy, so the idea of having per type lists is moot.

Instead, we should have a single list of all rules that belong to the
policy.

Also fixed a test that validated a single transaction to actually work
with multiple autonested transactions applied.

Partially-Implements: blueprint quantum-qos-api
Change-Id: Ia152b3ff385d2aa0cf40664ef039265b046b1d17