Fix hooks for dealing with member actions
This patch ensures that operations like add_router_interface
are correctly dispatched to the plugin.
It does so by introducing fairly ugly hacks in the
resource_identifier, and attribute_population pecan hooks.

This is however the best possible solution wiht the current
infrastructure which basically relies on hooks for request
dispatching.

Change-Id: I34880d1ce796cfd6f81d16fc1b9fd8387b5db056