Follow up with some cleanup for agent qos_driver
Removed error handling for missing rule type handler since the rule type
is already filtered thru supported types for the backend, so in case the
handler is really not present, that's a huge bug in the qos driver
extension and should not be handled gracefully.

Also fixed some grammar in docstrings.

Change-Id: Id157bd1e105051a583fea8e5107326289c551739
Partially-Implements: quantum-qos-api