Validate updated allocation pool before using it
The allocation pool specified during subnet-update is being
used in gateway validation checks before the allocation pool
is ever validated.  Errors indicating that the gateway IP is
invalid are returned when it is the allocation pool that is
actually in error.

Additional testing of the allocation pool is also being added
to the subnet-update unit tests.

Closes-Bug: #1479948
Change-Id: I854333d571bad2550ea5a0b93ff64ce44d0bd010