Introduce a separate RPC server
As the Pecan server only server REST requests over HTTP, this
patch introduces a new server implementing the RPC over AMQP
endpoints for agent/server communication.

However, the REST server does not yet have the ability to send
notifications to the RPC server or directly to the agents.
This patch simply adapts the ML2 plugin to run the RPC notifiers
only when initialized in the pecan server, so that notification
to agents can still be sent.

This patch therefore is tantamount to a poor man's
implementation of REST/RPC separation which will be iteratively
improved.

Change-Id: Ie471869d9b2793acdc412f13507038433f6a72c6