Implement TODO for version listing
Return a list of currently enabled API versions building
the response in the same way as it is done for the eventlet
WSGI API in order to ensure backward compatibility.

The only version listed is obviously 2.0, but the solution
devised in this patch allows version controller for self
declaring the version they implement.

This patch also replaces the use of _lookup when such method
is not strictly necessary.
To this aim, the minimum version of Pecan has been bumped to
1.0 as previouse releases did not support dots "." as valid
characters in URI path segments.

Change-Id: Ia748194b2c07a04a711d32ce8058a006e73621ef