Fix DBDuplicateEntry when creating port with fixed_ips on PD subnet
Creating a port on a prefix delegation enabled subnet with a fixed ip
list containing only the subnet id will cause a DBDuplicateEntry error.
This is because the subnet is not treated as an ipv6 auto address subnet
as it should be by _test_fixed_ips_for_port in ipam_non_pluggable_backend.py

All IPV6 PD subnets are auto address subnets, so the additional check was
unnecessary and incorrect in this case.

This patch changes the condition for appending the subnet id to the
fixed_ip_set so that PD subnets are treated the same as auto address
subnets. Also includes a unit test to catch this failure.

Change-Id: I8ad499467a09133258b1e9e40db736d4e1ae62f3
Closes-Bug: 1484379