Get rid of exception converter in db/api.py
The exception converter was necessary because the exceptions the
oslo db decorator looked for before were statically defined. The
retry decorator now accepts an exception_checker argument that takes
a function to call on exceptions to determine if they should be
caught.

This patch gets rid of the converted and replaces the one use case
with the new exception_checker argument.

Closes-Bug: #1485819
Change-Id: Ic619b03737cbf51276f87c4458ecc4183424731c