Process update_network in the openvswitch agent
This will allow ports with attributes related to the network to be
updated as necessary. Initially QoS extension which is able to
attach a network policy to the port.

Another approach would be sending updates to every single port
on a network, but that doesn't scale well for networks with lots
of ports.

Change-Id: Ie28297840b5702a920142af02dd17b10775d76ca
Partially-Implements: blueprint ml2-qos
Closes-Bug: 1486028