Stop logging STDOUT and STDERR on every shell out
Sometimes you can log too much. For example, logging a complete
iptables dump on every security group operation, OMG TMI. Doing this
during DSVM run results in a log file >7M compressed. To mitigate this
issue this commit switches the execute method to only log the command
and it's exit code on the success case. If there is a failure the
entire stdin, stdout and stderr are logged.

Change-Id: Iaf17297306dc752e666612033c805a528f078f2f