Fix AttributeError in _clean_updated_sg_member_conntrack_entries()
The conntrack work that recently merged introduced a bug
due to an incorrect 'if' statement, it was only detecting
when one variable was not set instead of both, which can
cause an exception.  This is currently causing jenkins failures
with a number of other changes.

Introduced in Change Id Ibfd2d6a11aa970ea9e5009f4c4b858544d8b7463

Change-Id: I2519fdceefc9255c21d8226cdeffec40a7d444f6
Closes-bug: #1488284