NSX-mh: use router_distributed flag
Even if the plugin technically already leverages this flag as it's
used in the l3_dvr_db mixin, the logic for the NSX-mh implies that
it is the backend that decides whether routers by default are
centralized or distributed.

This patches changes the logic for the default value for distributed
routers, ensuring the router_distributed flag is honoured.

(Removed from gate: Neutron failing all unit tests)


Applied from commit: 0649bb61f97f9a24dda525ec02a8f531914ae02c
Change-Id: I6cee010622d310477867b94b754aa2260b43b879
Closes-Bug: #1483920