Add tenant_id to flavor service profiles attributes
Neutron v2 base.py auto populates a 'tenant_id' attribute on
calls if the attribute is not passed.  This causes a POST
to create a flavor service binding to fail when verifying
attributes with:
Unrecognized attribute(s) 'tenant_id'

Solution is to add tenant_id as expected attribute in the
attribute map as done in other sub resources like QOS.

Fix unit test for non-keystone case.

Change-Id: Ic2bd1271f297fc10b49304ffd5fe617637e3d8f4
Closes-Bug: 1489197