create_router: Report the original exception
In case delete_router, which is an attempt to revert
the operation, failed, report the original exception
rather than the one from delete_router.  The former is
more likely useful for users.  Besides that, it's critical
to preserve the original exception for special exceptions
like RetryRequest.

Also, add some relevant unit tests.

Closes-Bug: #1490917
Change-Id: I1cef944115c1911d4b7e60fb4dff389b1911ddbb