Retry metadata request on connection refused error
This testcase may fail intermittently on 'Connection refused' error.
This could be due to the fact that the metadata proxy setup is not exactly
complete at the time the request is issued; in fact there is no
synchronization between the router being up and the metadata request being
issued, and clearly this may be the reason of accidental but seldom failures.

In order to rule out this possibility and stabilize the test, let's retry
on connection refused only. If we continue to fail, then the next step would
be to dump the content of iptables to figure out why the error occurs.

Closes-bug: #1461172

Conflicts:
	neutron/tests/functional/agent/test_l3_agent.py

Change-Id: I65a5bf4fbbcad6ba93a46d36cabe7844ff528d8d
(cherry picked from commit 4b2e6842f320405cd963f560bc06849b4b7bb1eb)