ovs: don't use ARP responder for IPv6 addresses
ARP does not support IPv6 addresses, so when we try to apply the flow, it
fails, with all other flows deferred for the same transaction. It results in
random flow breakages, depending on the order of the bad flow in the
transaction.

Conflicts:
	neutron/plugins/openvswitch/agent/ovs_neutron_agent.py
	neutron/tests/unit/plugins/ml2/drivers/openvswitch/agent/test_ovs_neutron_agent.py

Change-Id: I0ecf167653e5a7d0916e091e05050406a026a1e2
Co-Authored-By: Thomas Carroll <Thomas.Carroll@pnnl.gov>
Closes-Bug: #1477253
(cherry picked from commit a19e64c9d95781982d28113c667dbc90d0ea11eb)