Make OVS interface name hashing algorithm common and extend it
The OVS devcie name hashing algorithm shortens interface names that are too
long. To ensure uniqueness it takes use of a hashing algorithm.

Move this function to a common place where it can be shared between ml2
drivers and agents.

Extend the function to support interface name postfix in addition to the
prefix.

Change LOG level to info to help deployers figuring out the unhashed name
of an hashed inteface.

Adapt OVS and linuxbridge agent to use this common function instead of their
own implementation.

Change-Id: If7ee8240c8f60da3400365138832a5d9badf98b4
Closes-Bug: #1495960