Add --verbose to subset of cmds in neutron-db-manage
This commit updates neutron/db/migration/cli.py such that --verbose can
be passed to the following commands:

- current
- history
- branches

We also update tests in neutron/tests/unit/db/test_migration.py and add
a new test for 'neutron-db-manage branches' as that appears to be
untested.

Lastly, we add a brief description of the newly added capability to
doc/source/devref/alembic_migrations.rst.

Change-Id: I9fc136055b422f12a22c1365f52f17df53219820
Closes-Bug: #1488021