Remove early yields in _iter_hosts in dhcp agent
There were early yields in the DHCP agent to handling skipping
lease setup for certain IPv6 addresses. However, the setup for
name resolution in _output_addn_hosts_file depends on the same
function so certain IPv6 addresses weren't getting name resolution.

This patch adjusts the function to return a flag indicating that
DHCP isn't required for the port so the callers setting up reservations
can skip based on that and the name resolution setup can continue
as normal.

Change-Id: I81b4669eadaa9119e08c6a5e1d2a7b5959babdcc
Closes-Bug: #1498665