Kill HEADS file
Now that we forbid branch points as part of check_migration, there is no
technical need in maintaining the file. It only slows down developers that are
forced to resolve git conflicts after every new migration merged.

We remove the file as part of new 'revision' generation.

Now that the file is not present, we introspect the alembic scripts to
determine whether the subproject uses split migrations (presense of a branch
point indicates that).

Change-Id: Ic1ae0c3fc70be859919b89444dbfdcb1068bb1cd