Static routes not added to qrouter namespace for DVR
Today static routes are added to the SNAT namespace
for DVR routers. But they are not added to the qrouter
namespace.

Also while configuring the static routes to SNAT
namespace, the router is not checked for the existence
of the gateway.

When routes are added to a router without a gateway the
routes are only configured in the router namespace, but
when a gateway is set later, those routes have to be
populated in the snat_namespace as well.

This patch addresses the above mentioned issues.

Closes-Bug: #1499785
Closes-Bug: #1499787

Conflicts:
	neutron/agent/l3/dvr_edge_router.py
        neutron/tests/functional/agent/test_l3_agent.py
	neutron/tests/functional/agent/l3/framework.py
	neutron/tests/functional/agent/l3/test_dvr_router.py

Change-Id: I37e0d0d723fcc727faa09028045b776957c75a82
(cherry picked from commit 158f9eabe20824b2c91eaac795dad8b8a773611d)