Tag the alembic migration revisions for Liberty
Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.

So from now on we tag the final revision on every branch when we make
a milestone release. Update the cli to support the command:
  neutron-db-manage upgrade <milestone>
where <milestone> becomes an alias for all the revisions for a
milestone.

Closes-Bug: #1499033

Change-Id: I38623986dd574bec01fe147f9c6a747f3f512bb7
(cherry picked from commit 52236764a3a7c23ec24a86ccb84bc6737e2a4791)
Conflicts:
        neutron/db/migration/cli.py
	neutron/tests/unit/db/test_migration.py