Remove zombie pecan hook
The resource_identifier hooks has been removed as the Pecan
controllers are now able to identify by themselves the resource
they operate for.

However the unused hook module was never removed thus making it
'undeed'. This patch simply fires it a silver bullet thus ensuring
it will be gone forever.

Change-Id: I421c3ccffb6d01846db636c878869109198af7ad