Split the FIP Namespace delete in L3 agent for DVR
Right now we are seeing a race condition in the l3 agent
for DVR routers when a floatingip is deleted and created.

The agent tries to delete the floatingip namespace and
while it tries to delete there is another call to add a
namespace. There is a timing window in between these two
calls where sometimes the call to create a namespace succeeds
but, when tried to execute any commands in the namespace
it fails, since the namespace was deleted concurrently.

Since the fip namespace is associated with an external net
and each node has only one fip namespace for an external net,
we would like to only delete the fip namespace when the
external net is deleted.

The first step is to split the delete functionality into two.
The call to fip_ns.cleanup will only remove the dependency that
the fipnamespace has with the router namespace such as fpr and
rfp veth pairs.

The call to fip_ns.delete will actually delete the
the fip namespace and the fg device.

Partial-Bug: #1501873
(cherry picked from commit c874f6dadaa983ed4f338808907c8829a7f86031)
Change-Id: Ic94625d5a968f554af70c274b2b2c20ab64e2487