Simplify L3 HA scheduler tests
Stop mocking scheduling in the L3 HA scheduler base class,
and mock it explicitly in a couple of tests that need to control
scheduling manually. This allows us to remove a tiny bit of code,
but more importantly make the test and production code a little bit
more similar, and allow the base class to be reused in a more
generic manner.

Change-Id: I95790ba0605bcbf9b87199b99116aee4fca4d604