Added postcommit hooks in ovsvapp mech_driver
This patch-set added the postcommit methods for network and port
in ovsvapp mechanism driver which is required by ovsvapp driver.py
sitting in openstack/networking-vsphere repo.

This issue was addressed in 'master' (Mitaka) as part of review:
https://review.openstack.org/#/c/240127/

Backporting the relevant part of that fix here, since 
cherrypicking is not possible.

Closes-Bug: #1512328

Change-Id: I066891d83f005c7214a2496812ca506003d32c5f