Update default gateway in the fip namespace after subnet-update
The gateway does not get updated after a subnet-update --gateway
because the router is already subscribed to the fip namespace.
On a restart, the fip namespace has no subscribers, which causes
the gateway to be created. You shouldn't have to restart the L3
agent in order to update the gateway.

In order for the default gateway in the fip namespace to be updated
after a subnet-update, the underlying code which sets the default
gateway needs to be called even if the router is already subscribed to
the fip namespace.

The bit of code which is responsible for updating the default route has
been refactored into its own method, update_gateway_port, which is then
called after a 'subnet-update --gateway ...'. If the gateway has
actually changed, the default route in the fip namespace will be
updated.

Conflicts:
	neutron/tests/unit/agent/l3/test_dvr_fip_ns.py

Change-Id: I0552c4a50ae09a6606e88395f7545179d02733dd
Closes-Bug: #1512858
(cherry picked from commit a0f7153f9512ce313f0c8762889fe3ef6b455a32)