Use get_interface_bridge instead of get_bridge_for_tap_device
bridge_lib defines get_interface_bridge which returns the bridge on
which an interface is connected (or None) where LB L2-agent[1] returns
it only if the bridge is a bridge managed by LB L2-agent.

According to get_bridge_for_tap_device 1st usage:

  if bridge_lib.is_bridged_interface(interface):
     bridge = self.get_bridge_for_tap_device(interface)
     bridge.delif(interface)

The code crashs if interface is owned by a non-LB-agent bridge because
the condition is verified but bridge is None.

[1] neutron.plugins.ml2.drivers.linuxbridge.agent.linuxbridge_neutron_agent

Partial-Bug: #1514548
Change-Id: If6dba1197b3f08ecc0ac01fcf085c8268f81169f