Fix dvr_local_router.floating_ip_added_dist failure after agent restart
If DVR is enabled and an l3-agent is restarted, associating a floatingip
to an existing vm on the compute node where the l3-agent was restarted
will fail.  This commit fixes the floating_ip_added_dist method in
dvr_local_router.py so that self.rtr_fip_subnet will be created if
it is 'None' prior to the attempt to call the get_name() method on
it.

Closes-Bug: #1515341
(cherry picked from commit da3de52c507544c84463a037acac3b2b24a37fe0)

Conflicts:
	neutron/tests/unit/agent/l3/test_dvr_local_router.py

Change-Id: I860ee3522275f62f393e66853e53ef387598983d