Elaborate how priorities are assigned to blueprints
Prioritizing is hard, and prioritizing in open source, let
alone Openstack, is even harder.

For this reason, this proposal attempts to clarify what priority
mean, how we prioritize in Neutron, in a way that's objective,
consistent, and reproducible. To start with, and make things slightly
simpler, we reduce the number of priority buckets we can choose from
to two: High and Low.

The way priorities are assigned is then based on a set of clearly defined
criteria: high priority work items are items that all core folks
collectively should care about because any failure in achieving tangible
result make them look bad. Low work items are great to have, but if they
slipped, it would be a shame but not the end of the world.

Keep on reading to learn more.

Change-Id: I0f4ba622eb3d8e02b0c3204967fb8eae22db09c8