Send 50% less debug information when executing cmd
The current code prints out the command it runs twice: once in
'create_process' or 'execute_rootwrap_daemon', and once after the
command has finished executing and a return code is returned. Since this
is synchronous, there is no need to print the command twice in a row, so
this patch deletes the latter one, leaving only the exit code to be
printed.

This patch also refactors the surrounding code to make it more readable.

Change-Id: I22289f838dbc3956a777a78c882e4d5f1e725d0e