Don't add default route to HA router if there is no gateway ip
When adding an external network with no gateway ip to a HA router,
l3 agent will report exception. The exception comes from the code
of adding default route to HA router. However, if there is no
gateway ip in the external network, there is no need to add such
route.

Change-Id: I41d6a292c903758f408d3d93a64dca7adeeb5769
Closes-Bug: #1515209