Remove misplaced copyright attribution
the module neutron.api.extension has a copyright attribute to a user
who never contributed a line to this module. It was probably copied
over in the initial commit for the module.

This commit is probably quite pointless; however it removes a little
oddity and does no harm to anyone.

Change-Id: I985bf2cb0538a795464dfe783108f218f821d762