Clarify how we milestones are assigned
Milestone assignment is another mumbo jumbo effort in open source.
Artificially setting milestones implies that someone can reliably
predict the future when no-one is really in full control.

For this reason let's make clear that we optmistically target the
current milestone for work that is supposed to start asap, and
complete sooner rather than later. Rolling over until the work is
complete is the natural course of action.

Dashboards [1] then capture the entire workload (BP and RFE) for
the entire release cycle, and that's helpful to provide to overall
view.

[1] https://launchpad.net/neutron/+milestone/<release>-?

Change-Id: Idb2d84ba683d2c8f1460e7bf0ff76d457cf42bce