Fix incomprehensible releasenotes
Fix incorrect usage of RST for releasenotes, and also cleanup how we use
reno for these releasenotes. When they were orginally pushed, how reno
formatted things wasn't clear to the original author of these releasnotes,
resulting in suboptimal releasenotes. This cleans things up in a more
presentable format for all consumers of the releasenotes.

This commit is made directly to stable/liberty because these releasenotes
do not exist in master.

Closes-Bug: #1519912
Closes-Bug: #1519915

Change-Id: Iea07d53067f5ad7a5282982cb12a834c786a75b9
Signed-off-by: Kyle Mestery <mestery@mestery.com>